Removed old classes for NetChanges and Configuration
