--- deployments/helm/onos-config/files/configs/configStore-sample.json
@@ -1,126 +1,0 @@
-{
-  "Version": "1.0.0",
-  "Storetype": "config",
-  "Store": {
-    "Device1-1.0.0": {
-      "Name": "Device1-1.0.0",
-      "Device": "Device1",
-      "Version": "1.0.0",
-      "Type": "TestDevice",
-      "Created": "2019-05-09T16:24:17Z",
-      "Updated": "2019-05-09T16:24:17Z",
-      "Changes": [
-        "je06bYEn4tzIBDUTNcBOhHQYhg0=",
-        "P9dRXqy6IPp2KxeWxGvV0lRQIus=",
-        "Lj+MiMR60C3BLIoEbsp3R4rPzcg="
-      ]
-    },
-    "Device2-1.0.0": {
-      "Name": "Device2-1.0.0",
-      "Device": "Device2",
-      "Version": "1.0.0",
-      "Type": "TestDevice",
-      "Created": "2019-05-09T16:24:17.815298386Z",
-      "Updated": "2019-05-09T16:24:17.815300015Z",
-      "Changes": [
-        "je06bYEn4tzIBDUTNcBOhHQYhg0=",
-        "P9dRXqy6IPp2KxeWxGvV0lRQIus=",
-        "Wv2Ba23stoAcdD9PkaPUih2HYtY="
-      ]
-    },
-    "Device2-2.0.0": {
-      "Name": "Device2-2.0.0",
-      "Device": "Device2",
-      "Version": "2.0.0",
-      "Type": "TestDevice",
-      "Created": "2019-05-09T17:00:00Z",
-      "Updated": "2019-05-09T17:00:00Z",
-      "Changes": [
-        "je06bYEn4tzIBDUTNcBOhHQYhg0="
-      ]
-    },
-    "localhost-1-1.0.0": {
-      "Name": "localhost-1-1.0.0",
-      "Device": "localhost-1",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-21T08:43:17Z",
-      "Updated": "2019-05-21T08:43:17Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "/wbjsT1Pl4OKLUxHy+j8UBDySvk="
-      ]
-    },
-    "localhost-2-1.0.0": {
-      "Name": "localhost-2-1.0.0",
-      "Device": "localhost-2",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-21T08:43:18Z",
-      "Updated": "2019-05-21T08:43:18Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "HI1vkKfDhdk4g4LdPcr4Rd0DIeI="
-      ]
-    },
-    "localhost-3-1.0.0": {
-      "Name": "localhost-3-1.0.0",
-      "Device": "localhost-3",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-21T08:43:19Z",
-      "Updated": "2019-05-21T08:43:19Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "JP9oUsUllBpdWIR0XvyiUWxiv6c="
-      ]
-    },
-    "device-1-device-simulator-1.0.0": {
-      "Name": "device-1-device-simulator-1.0.0",
-      "Device": "device-1-device-simulator",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-30T18:03:17Z",
-      "Updated": "2019-05-30T18:03:17Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "QWEE7VidGo/Z4IiSTbK1wmwDtdo="
-      ]
-    },
-    "device-2-device-simulator-1.0.0": {
-      "Name": "device-2-device-simulator-1.0.0",
-      "Device": "device-2-device-simulator",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-30T18:03:17Z",
-      "Updated": "2019-05-30T18:03:17Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "KRjTNltrz0idwiBupo2qftfCN+A="
-      ]
-    },
-    "device-3-device-simulator-1.0.0": {
-      "Name": "device-3-device-simulator-1.0.0",
-      "Device": "device-3-device-simulator",
-      "Version": "1.0.0",
-      "Type": "Devicesim",
-      "Created": "2019-05-30T18:03:17Z",
-      "Updated": "2019-05-30T18:03:17Z",
-      "Changes": [
-        "mSr0HBYynI8r7bcpEcOXz4bL4aw=",
-        "Jcx94M+jOyu4q5A3zycqgMnAdi8="
-      ]
-    },
-    "stratum-sim-1-1.0.0": {
-      "Name": "stratum-sim-1-1.0.0",
-      "Device": "stratum-sim-1",
-      "Version": "1.0.0",
-      "Type": "Stratum",
-      "Created": "2019-06-05T10:03:17Z",
-      "Updated": "2019-06-05T10:03:17Z",
-      "Changes": [
-        "QANqS/52DozB/OUKENoVcvOkUZs="
-      ]
-    }
-  }
-}
