--- deployments/helm/onos-config/files/configs/networkStore-sample.json
@@ -1,14 +1,0 @@
-{
-  "Version": "1.0.0",
-  "Storetype": "network",
-  "Store": [
-    {
-      "Name": "Removed txout-1 from Device 2",
-      "Created": "2019-05-10T00:00:00Z",
-      "User": "User1",
-      "ConfigurationChanges": {
-        "Device1-1.0.0": "FGrHfVhk8rwhJ03U5TwxFqPSB/8="
-      }
-    }
-  ]
-}
