--- deployments/helm/onos-config/templates/configmap.yaml
@@ -1,29 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{ template "fullname" . }}-config
-  labels:
-    app: {{ template "fullname" . }}
-    chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
-    release: "{{ .Release.Name }}"
-    heritage: "{{ .Release.Service }}"
-data:
-  deviceStore.json: |
-    {
-      "Version": "1.0.0",
-      "Storetype": "device",
-      "Store": {
-        {{- range $i, $device := .Values.devices }}
-        {{ if $i }},{{ end}}{{ printf "%s" $device | quote }}: {
-          "ID": {{ printf "%s" $device | quote }},
-          "Addr": {{ printf "%s:10161" $device | quote }},
-          "Timeout": 5,
-          "SoftwareVersion": "1.0.0"
-        }
-        {{- end }}
-      }
-    }
-  {{ $root := . }}
-  {{ range $path, $bytes := .Files.Glob "files/configs/*.json" }}
-  {{ base $path }}: '{{ $root.Files.Get $path }}'
-  {{ end }}
