--- pkg/store/networkconfiguration.go
@@ -1,56 +1,0 @@
-// Copyright 2019-present Open Networking Foundation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package store
-
-import (
-	"fmt"
-	"github.com/onosproject/onos-config/pkg/store/change"
-	"regexp"
-	"time"
-)
-
-// NetworkConfiguration is a model of a network defined configuration
-// change. It is a combination of changes to devices (configurations)
-// Deprecated: NetworkConfiguration is a legacy implementation of an internal NetworkConfiguration
-type NetworkConfiguration struct {
-	// Deprecated
-	Name string
-	// Deprecated
-	Created time.Time
-	// Deprecated
-	User string
-	// Deprecated
-	ConfigurationChanges map[ConfigName]change.ID
-}
-
-// NewNetworkConfiguration creates a NetworkConfiguration object
-// If name is left blank then a name will be generated by a hash algorithm
-// Deprecated: NewNetworkConfiguration is a method for creating internal NetworkConfiguration
-func NewNetworkConfiguration(name string, user string, changes map[ConfigName]change.ID) (*NetworkConfiguration, error) {
-	r1 := regexp.MustCompile(`[a-zA-Z0-9\-_]+`)
-	match := r1.FindString(name)
-	if name == "" {
-		return nil, fmt.Errorf("Empty name not allowed")
-	} else if name != match {
-		return nil, fmt.Errorf("Error in name %s", name)
-	}
-
-	return &NetworkConfiguration{
-		Name:                 name,
-		Created:              time.Now(),
-		User:                 user,
-		ConfigurationChanges: changes,
-	}, nil
-}
